refactor(api): paginate device/user queries and split OpenAPI specs
- Replace 0-based @ParameterObject pagination with explicit 1-based page/size
  params and a shared PageRequestBuilder across user/devices/admin endpoints
- Return Page from user-device repository/service and align cache keys with
  annotation-driven eviction
- Make OTA dispatch transactional and evict affected caches after fan-out
- Collapse V2-V9 flyway migrations into the baseline V1 schema
- Omit null error/code fields from 2xx ApiResponse payloads
- Document conditional TimeWindowRequest fields in the OpenAPI schema
- Use int32 for device displayOrder in API DTOs
- Change device secret retrieval from GET to POST to avoid credential logging
- Split /v1/internal endpoints into a dedicated internal OpenAPI group while
  keeping them out of the public spec
- Add nanomq healthcheck to docker-compose

+++ src/main/java/dev/ivfrost/hydro_backend/devices/DeviceResponse.java
@@ -21,6 +21,6 @@ public record DeviceResponse(
     @NotNull Instant updatedAt,
     @NotNull Instant linkedAt,
     @NotNull Long userId,
-    @NotNull Long displayOrder) {
+    @NotNull Integer displayOrder) {
 
 }
